+++ eggs/Naaya/Products/NaayaCore/EmailTool/EmailValidator.py
@@ -1,20 +1,34 @@
 import os
+import re
 import time
 import logging
-LOG = logging.getLogger(__name__)
 from threading import Thread, Lock, Semaphore
 from Products.NaayaCore.EmailTool.SharedSet import SharedSet, Full, Empty
-
-from validate_email import validate_email
-
 import transaction
+
+
+LOG = logging.getLogger(__name__)
+
+
+def validate_email(email):
+    pattern = (r"(?:^|\s)[-a-z-A-Z0-9_.']+@(?:[-a-z-A-Z0-9]+\.)+[a-z-A-Z]"
+               r"{2,63}(?:\s|$)")
+    return bool(re.match(pattern, email))
+
 
 class EmailValidator(object):
     """Runs mail addresses through validate_email lib using detached threads
     to parallelize this action and not keep the curent op waiting in I/O
     Make sure to use bind() befaore enqueue() so that the result can be stored
     asynchronously. outputObj must have a dict like property named by string
-    storage_name."""
+    storage_name.
+
+    It was originally developed to work with the validate_email library
+    that tried to communicate with each email server to actually check if the
+    email address exists there, a process that needed to be done
+    asynchronously because of its duration, but since a few years email servers
+    don't answer to such requests anymore (spam protection), so the library was
+    removed, but the semaphore procedure was not (yet) removed """
 
     # Note that while we avoid most of the duplicate validations,
     # they may still occur when one thread spends time waiting for validation,
@@ -30,7 +44,7 @@ class EmailValidator(object):
         VERIFY_EMAIL_BADADDRESS_TTL = (24 * 60 * 60)
         VERIFY_EMAIL_GOODADDRESS_TTL = (30 * 24 * 60 * 60)
     THREAD_IDLE_SEC = 60
-    VALIDATION_ATTEMPTS = 3 # number of consecutive attempt to avoid false negatives
+    VALIDATION_ATTEMPTS = 3
 
     def __init__(self, storage_name, maxWorkers=10, maxPoolSize=4000):
         self._outputObj = None
@@ -49,11 +63,13 @@ class EmailValidator(object):
 
     def _setupWorkers(self):
         while self._workerCountSemaphore.acquire(False):
-            name = "emailValidationThread_%d"%self._seq
+            name = "emailValidationThread_%d" % self._seq
             self._workersLock.acquire()
-            self._workers[name] = {'th': Thread(
-                target=EmailValidator._worker, name=name, args=(self,name)),
-                                  'running': None}
+            self._workers[name] = {
+                'th': Thread(
+                    target=EmailValidator._worker, name=name,
+                    args=(self, name)),
+                'running': None}
             self._workersLock.release()
             self._seq += 1
 
@@ -70,29 +86,38 @@ class EmailValidator(object):
 
     def bind(self, outputObj):
         if not hasattr(outputObj, self._storage_name):
-            raise RuntimeError("{} instance should be bound to an object containing {}"
-                " PersistentMapping".format(self.__class__.__name__, self._storage_name))
+            raise RuntimeError(
+                "{} instance should be bound to an object containing {}"
+                " PersistentMapping".format
+                (self.__class__.__name__, self._storage_name))
         self._outputObj = outputObj
 
     def validate_from_cache(self, email):
         now = time.time()
-        check_value, check_ts = getattr(self._outputObj, self._storage_name).get(email, (None, None));
-        if ( check_value is None
-            or (check_value is not True and check_ts < now - self.VERIFY_EMAIL_BADADDRESS_TTL)
-            or (check_value is True and check_ts < now - self.VERIFY_EMAIL_GOODADDRESS_TTL) ):
+        check_value, check_ts = getattr(self._outputObj,
+                                        self._storage_name).get(email,
+                                                                (None, None))
+        if (check_value is None or
+                (check_value is not True and
+                 check_ts < now - self.VERIFY_EMAIL_BADADDRESS_TTL) or
+            (check_value is True and
+                check_ts < now - self.VERIFY_EMAIL_GOODADDRESS_TTL)):
             return None
         return check_value
 
-
     def enqueue(self, email):
         if not self._outputObj:
-            raise RuntimeError("{} instance should be bound to an object containing {}"
-                " PersistentMapping".format(self.__class__.__name__, self._storage_name))
+            raise RuntimeError(
+                "{} instance should be bound to an object containing {}"
+                " PersistentMapping".format(self.__class__.__name__,
+                                            self._storage_name))
         self._setupWorkers()
         try:
             self._inQ.put_nowait(email)
         except Full:
-            LOG.warn("input validate mail queue full. %s will not be resolved now" % email)
+            LOG.warn(
+                "input validate mail queue full. %s will not be resolved now" %
+                email)
 
     def _worker(self, name):
         LOG.debug("new thread started: %s" % name)
@@ -106,21 +131,22 @@ class EmailValidator(object):
                 for i in range(self.VALIDATION_ATTEMPTS):
                     # long I/O bound operation
                     try:
-                        check_value = validate_email(email, verify=True,
-                                                     verbose=True)
-                    except:
+                        check_value = validate_email(email)
+                    except AttributeError:
                         check_value = False
                     if check_value:
                         break
 
                 self._outputObjLock.acquire()
                 now = time.time()
-                getattr(self._outputObj, self._storage_name)[email] = (check_value, now)
+                getattr(self._outputObj,
+                        self._storage_name)[email] = (check_value, now)
                 transaction.commit()
                 self._outputObjLock.release()
             self._inQ.task_done()
         self._workersLock.acquire()
         self._workers[name]['running'] = False
         self._workersLock.release()
-        LOG.debug("thread exits after %d seconds idle: %s" % (self.THREAD_IDLE_SEC, name))
+        LOG.debug("thread exits after %d seconds idle: %s" %
+                  (self.THREAD_IDLE_SEC, name))
         self._workerCountSemaphore.release()